backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain_openai
openai
nltk
rapidfuzz
requests
python-multipart
